chore(test): Move run_fabulous_commands_with_logging to helpers
